ErrorContainer and Loading moved into common/sub-pages.
Maybe we should name sub-pages -> containers?
