+++ src/components/common/sub-pages/Loading.tsx
@@ -4,7 +4,7 @@ import loadingIcon from '@iconify-icons/mdi/loading';
 
 
 const Loading = (): JSX.Element => (
-  <div className="container" role="status">
+  <div className="container m-auto" role="status">
      <svg className="animate-spin m-auto" width="50" height="50">
       <Icon icon={loadingIcon} width="50" height="50"/>
     </svg>
